rename hive to vertex
