+++ app/(auth)/vertex/components/icons/langgraph.tsx
