+++ app/(auth)/vertex/components/thread/agent-inbox/components/tool-call-table.tsx
