+++ app/(auth)/vertex/components/thread/agent-inbox/index.tsx
@@ -2,7 +2,7 @@ import { StateView } from "./components/state-view";
 import { ThreadActionsView } from "./components/thread-actions-view";
 import { useState } from "react";
 import { HumanInterrupt } from "@langchain/langgraph/prebuilt";
-import { useStreamContext } from "@/app/(auth)/hive/providers/Stream";
+import { useStreamContext } from "@/app/(auth)/vertex/providers/Stream";
 
 interface ThreadViewProps {
   interrupt: HumanInterrupt | HumanInterrupt[];
