+++ app/(auth)/vertex/components/thread/agent-inbox/utils.ts
