+++ app/(auth)/vertex/components/thread/messages/generic-interrupt.tsx
