+++ app/(auth)/vertex/components/thread/messages/tool-calls.tsx
