+++ app/(auth)/vertex/components/thread/referenced-text-display.tsx
