+++ app/(auth)/vertex/components/thread/selectable-content.tsx
