+++ app/(auth)/vertex/components/thread/selection-action-button.tsx
