+++ app/(auth)/vertex/components/thread/syntax-highlighter.tsx
