+++ app/(auth)/vertex/components/thread/utils.ts
