+++ app/(auth)/vertex/components/ui/avatar.tsx
