+++ app/(auth)/vertex/hooks/useTextSelection.tsx
