+++ app/(auth)/vertex/lib/api-key.tsx
