+++ app/(auth)/vertex/lib/ensure-tool-responses.ts
