+++ app/(auth)/vertex/lib/utils.ts
