+++ app/(auth)/vertex/providers/ReferencedText.tsx
