+++ app/(auth)/vertex/providers/client.ts
